refactor(library management): renaming and restructuring
